requirements.txt
torch==2.7.1
gymnasium[box2d]==1.1.1
streamlit==1.45.1
pillow==11.2.1
numpy==2.3.0










